project_files/utils/requirements.txt
absl-py==2.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-jwt-auth==0.5.0
fonttools==4.56.0
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
immutabledict==4.2.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.22
langchain-community==0.3.20
langchain-core==0.3.49
langchain-openai==0.3.11
langchain-pinecone==0.2.3
langchain-tests==0.3.17
langchain-text-splitters==0.3.7
langsmith==0.3.22
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
multidict==6.3.1
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.70.0
openpyxl==3.1.5
orjson==3.10.16
ortools==9.12.4544
packaging==24.2
pandas==2.2.3
pillow==11.2.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
propcache==0.3.1
protobuf==5.29.4
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pydub==0.25.1
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SoundCard==0.4.4
SQLAlchemy==2.0.40
syrupy==4.9.1
tenacity==9.0.0
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
yarl==1.18.3
zstandard==0.23.0